Many enhancements

+++ src/Controller/TaskController.php
@@ -5,37 +5,83 @@ namespace App\Controller;
 use App\Entity\Task;
 use App\Entity\User;
 use App\Form\TaskType;
+use App\Paginator\PaginatorService;
+use App\Repository\TaskRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Uid\Uuid;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
 class TaskController extends AbstractController
 {
     public function __construct(
         private readonly ManagerRegistry $managerRegistry,
-        private readonly TranslatorInterface $translator
+        private readonly TaskRepository $taskRepository,
+        private readonly TranslatorInterface $translator,
+        private readonly PaginatorService $paginatorService
     ) {
     }
 
     #[Route(path: '/tasks/todo', name: 'task_list_todo')]
-    public function todoList(): Response
+    public function todoList(Request $request): Response
     {
+        $paginationError = $this->paginatorService->create(
+            $this->taskRepository,
+            $request,
+            'task_list_todo'
+        );
+        if ($paginationError) {
+            $this->addFlash('error', $paginationError['message']);
+            $response = $this->render('task/list_todo.html.twig', [
+                'tasks' => null,
+            ]);
+            $response->setStatusCode(intval($paginationError['code']));
+
+            return $response;
+        }
+
         return $this->render('task/list_todo.html.twig', [
-            'tasks' => $this->managerRegistry->getRepository(Task::class)->findBy(['isDone' => false])
+            'tasks' => $this->paginatorService->getData(),
+            'currentPage' => $this->paginatorService->getCurrentPage(),
+            'firstPage' => $this->paginatorService->getUrlFirstPage(),
+            'lastPage' => $this->paginatorService->getUrlLastPage(),
+            'nextPage' => $this->paginatorService->getUrlNextPage(),
+            'previousPage' => $this->paginatorService->getUrlPreviousPage(),
         ]);
     }
 
     #[Route(path: '/tasks/finished', name: 'task_list_finished')]
-    public function finishedList(): Response
+    public function finishedList(Request $request): Response
     {
+        $paginationError = $this->paginatorService->create(
+            $this->taskRepository,
+            $request,
+            'task_list_finished'
+        );
+        if ($paginationError) {
+            $this->addFlash('error', $paginationError['message']);
+            $response = $this->render('task/list_finished.html.twig', [
+                'tasks' => null,
+            ]);
+            $response->setStatusCode(intval($paginationError['code']));
+
+            return $response;
+        }
+
         return $this->render(
             'task/list_finished.html.twig',
             [
-            'tasks' => $this->managerRegistry->getRepository(Task::class)->findBy(['isDone' => true])
-        ]);
+                'tasks' => $this->paginatorService->getData(),
+                'currentPage' => $this->paginatorService->getCurrentPage(),
+                'firstPage' => $this->paginatorService->getUrlFirstPage(),
+                'lastPage' => $this->paginatorService->getUrlLastPage(),
+                'nextPage' => $this->paginatorService->getUrlNextPage(),
+                'previousPage' => $this->paginatorService->getUrlPreviousPage(),
+            ]
+        );
     }
 
     #[Route(path: '/tasks/create', name: 'task_create')]
@@ -48,6 +94,7 @@ class TaskController extends AbstractController
             /** @var User $user */
             $user = $this->getUser();
             $task->setOwner($user);
+            $task->setUuid(Uuid::v4());
             $em = $this->managerRegistry->getManager();
             $em->persist($task);
             $em->flush();
@@ -59,9 +106,14 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    #[Route(path: '/tasks/{id}/edit', name: 'task_edit')]
+    #[Route(path: '/tasks/{uuid}/edit', name: 'task_edit')]
     public function edit(Task $task, Request $request): Response
     {
+        if (!$this->isGranted('TASK_EDIT', $task)) {
+            $this->addFlash('error', $this->translator->trans('app.flashes.task.user_not_authorized_to_edit'));
+
+            return $this->redirect($request->headers->get('referer'));
+        }
         $form = $this->createForm(TaskType::class, $task);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
@@ -77,7 +129,7 @@ class TaskController extends AbstractController
         ]);
     }
 
-    #[Route(path: '/tasks/{id}/toggle', name: 'task_toggle')]
+    #[Route(path: '/tasks/{uuid}/toggle', name: 'task_toggle')]
     public function toggleTask(Task $task): Response
     {
         $task->toggle(!$task->isDone());
@@ -88,13 +140,19 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list_todo');
     }
 
-    #[Route(path: '/tasks/{id}/delete', name: 'task_delete')]
+    #[Route(path: '/tasks/{uuid}/delete', name: 'task_delete')]
     public function deleteTask(Request $request, Task $task): Response
     {
+        if (!$this->isGranted('TASK_DELETE', $task)) {
+            $this->addFlash('error', $this->translator->trans('app.flashes.task.user_not_authorized_to_delete'));
+
+            return $this->redirect($request->headers->get('referer'));
+        }
         $em = $this->managerRegistry->getManager();
         $em->remove($task);
         $em->flush();
         $this->addFlash('success', $this->translator->trans('app.flashes.task.deleted'));
+
         return $this->redirect($request->headers->get('referer'));
     }
 }
